Adds SimpleFormattedText type
SimpleFormattedText is useful for presenting textual data where some
basic formatting is needed (HTML escaping and such) but the full
behavior of LCG StructuredText is not wanted.

Warning: This patch is a hack. SimpleFormattedText should be
implemented in LCG and independent of format. Currently, it
is exporting HTML. Also performance of the formatter could be
improved by doing everything in a single pass.

This patch requires a recent LCG due to dependency on the
new lcg.HtmlContent class.