Fix saving user configuration broken by commit 93d9b7edf0e3f59f8040547460c100ef76aa7bdf
Previous implementation of `Configuration.Option.changed()' returned True if
the current value was different from the initial value, while the
implementetion introduced by the named commit returns True if `set_value()' was
called, regardless whether the actual value differs or not.  This change breaks
saving user configuration, because it relied on the previous behavior.  This
fix doesn't change the behavior of `Configuration.Option.changed()', but rather
implements what is needed just for its own use.