Handle runtime filters for all filtered fields in web forms
The web form code checked for DataEnumerator to activate runtime filters for
given field, but this doesn't work for Array fields (which have the enumerator
in the inner type) and seems too limiting (other enumerators may also have
runtime filters).  The new solution leaves the decision up to the
PresentedRow.runtime_filter() method.  If it returns something else than None,
the field is activated for runtime filtering.