Rename the specification parameter 'initial_folding' to just 'folding'
All specification parameters define the initial form state so it is not
necessary to explicitly mark them as initial and as the others are not marked
as such, it is confusing to mark some of them.  Also with profiles, id would be
even more confusing.  The change is backwards compatible, so 'initial_folding'
defined in specifications is still recognized.