Use create ChecklistField links also on AJAX form updates
The links were correctly rendered only in the python code generating the
initial form.  But the ChecklistField checkboxes may be completely reloaded
within AJAX form updates so the JavaScript code needs to cake care of the same
thing.  This is a good example why it migh be a good idea to implement form
field rendering within JavaScript only, but thsi would be a big change.