Simplify value serialization by type-dependent primitive value conversions
Values of some pytis data types, such as DateTime, are not expressed internally
by basic python types.  This makes it problematic to serialize them to JSON or
other sililar formats.  This change adds a new method Type.primitive_value()
which is guaranteed to return a basic Python value.  Use this method if you
need seraalizable values.