Make webform work correctly in multiform sideforms
The hack previously used in the Browser initialization (embedding in the parent
window if the handle of self returns 0) made the browser panel behave
incorrectly in sizers.  It was actually really wrong to embed the browser into
its panel's parent.  The problem was in fact caused by the wx.Notebook, which
was not swithed to the browser tab in the time of its initialization.  This
commit fixes that but it may be quite a sensitive change in respect to delayed
form initialization.