Avoid unwanted removal of invalid profiles
Deletion of invalid profiles is now not queried on form startup but the
profiles are normally present in the menu and marked as invalid.  The question
about profile deletion is only displayed after the profile is selected which is
much less confusing for the user as he knows what he is doing (he doesn't
expect it on form startup).  The default answer is to keep the profile to
further avoid unwanted removals.