Drop support for non-database profile manager
The planned changes will require storing profile data in a structure of
database tables and implementation of the same behavior in python structures
would be inadequately demanding so it is easier to just drop this option.
Every pytis application will now require the database objects defined in
db/db_pytis_config.py to exist.

This commit also factors out some common profile manager code to be used by all
form data managers which are planned to be implemented.