Draw rectangle around current grid row rather than missusing selection
The selection makes the cell background invisible, which is unpleasant and is
also a reason for not displaying the selection in sideforms.  This, however
makes it impossible to distinguish the current sideform row.  It is also really
a missuse as the selection has a differnt purpose and mixing these two uses has
always been problematic.