Handle clipboard events through GTK
As GTK (unlike wx) makes it possible to monitor clipboard events, we now use it
to propagate local clipboard changes to the remote windows service (if
available).  It simplifies the current clipboard handling as it is only
important to copy text into the local clipboard it will get propagated
automatically, so it requires no special handling of copying within the
application (only pasting must be still handled specifically to distinguish the
source of the text).  It solves copying from the Webkit browser as well as any
other place in the application transparently.