Store common and form specific profile parameters separately
The set of profiles and their filters is now shared by all forms for given
specification.  Only form specific parameters, such as displayed columns, their
widths etc. are now stored separately for each form class.  This required
profile validation and all related logic to be moved from FormProfile into the
profile manager.  It actually cleaned up a lot of things as all profiles are
now of the same class, which is now immutable and the FormProfile class was
removed.