Image, Video, Audio and Resource dialogs in CKEditor
The above attachment types are now handled by separate
dialogs with a common base. All attachment types
with the exception of the generic resource incorporate
a preview.

Audio and video attachments are included in the HTML
source only as links to the target files. The possible
replacement of these semantic elements with more advanced
playback mechanisms (such as a Flash player or HTML5
elements) is left to additional javascript functions
not included in this patch.