Exercise dialog in CKEditor
The exercise dialog now allows for definition of the whole exercise
structure as described by the Eurochance manual. The only exception is
the insertion of audio pieces for dictation, which is currently
missing.

Additionally, this patch contains a slight but needed reorganization
of some code and naming of helper functions.