Changes to make database specifications closer to current HAS database schema
The changes are not necessarily correct from the application point of view.
Strong preference towards retaining current actual database schema was applied
as this is what is actually used and (mostly) working.  It may be worth to
review this commit and to revert regressions, together with applying them to
the database.