Ignore certain util/_test.py lines for coverage
First case is suppressing error during what is basically a test
tear-down, i.e. is not part of any logic verification

Second case is if __name__ == '__main__', which is a branch not taken
when running tests from the main lib/_test.py file. (Note that we do not
want to disable "if __name__ == '__main__'" globally in .coveragerc
since in som user-facing tools like gensql we are in principle
interested in coverage of code there.)