Fix ltree sorting after previous commit
The commit "Get rid of some clutter in DB column type initializations"
was first applied as 74a7724, but was later reverted by aad4d39, because
it introduced a regression resulting in ltree sorting not working
correctly in certain cases (sorting 'pytis_view_user_menu' by
'position').  This commit was re-applied as 1fc4ba9 and this commit
fixes the regression.