Work around invisible selection when bgcolor set
When a wxTextCtrl with an active selection looses focus, the selected
text becomes invisible when the field's background color is set.  The
selection color of unfocused field has the same color as the field
background color (for example light yellow by default for invalid field)
and the text color is white, so the contrast is so minimal that the text
is nearly invisible.