Fix ShowForm tab switching with CheckBox fields
It was impossible to switch tabs in a ShowForm when a checkbox field
appeared as the first field on the first tab in layouts with a TabGroup.

This fixes the problem as well as improves readability of read only
forms where we don't want to gray out ineditable fields (all fields are
ineditable in a read only form).