Aviod error on casting side form prefill values
We can not cast all values to their row types here, because some fields
may not exist in the row.  This happens when side form prefill (defined
in Binding specification) contains values for another form than for
itself when it redirects new record insertion to the other form through
on_new_record.