Toggle visible columns in dialog if too many
When there are too many columns, it is not very practical to toggle them
through the menu.  It is also not practical when you need to toggle
multiple columns at once.  Thus we use a separate dialog with checkboxes
including also column descriptions to simplyfy setting up the visible
columns.