Disable side form IN filters for binding conditon
Binding condition is dynamic (depends on main form row) and thus can not
be passed to the IN operator.  Maybe we could at least enable defining
static conditions (applied together with binding column condition) in
Binding specification.  It would be possible to pass such conditions to
the in operator and thus enable the IN operator in this case, but it is
probably quite a rare case not worth the trouble.