Fix `gensql` failures being unnoticed during CI
In shell pipe compound command, only the exit status of the last process
is the exit status of the whole compound command. So in case the last
command in the pipe exits with 0 status, the whole compound pipe command
will exit with 0 status (i.e. fine) even if previous commands in the
pipe errored with non-zero status. See
e.g. http://unix.stackexchange.com/questions/14270/get-exit-status-of-process-thats-piped-to-another

This could cause `gensql` errors to be unnoticed by the CI: `psql` did
not mind, as it got fed valid SQL and it cannot know that the SQL is
incomplete from the app logic point of view.

We now put `gensql` output into a file, and do not pipe it to `psql`.