Work around gsql failure on trigger creation
This fixes "AttributeError: 'NoneType' object has no attribute
'sqlalchemy_type'" when generating SQL for the 'CREATE TRIGGER'
statement (spotted in wiking.dbdefs, but probably occuring generally).
This is actually a quick fix without deep understanding the source of
the problem and the point of regression from the previously working
state.  It just seems to work...