Allow NOT in saved filters
The operator NOT is normally not needed as the filter dialogs don't
create conditions containing NOT.  However NOT may be used in system
profiles defined in specifications (typically with NOT IN) which may be
customized by users (such as ading/removing columns) and saved.  This
commit fixes the error when such customized profiles are loaded.