New configuration options use_dmp_roles and use_dmp_rights
Using these options it is possible to have all dmp tables
in the database but not to use them at all. It is also possible
to use only dmp roles, although the management of access rights
still remain on the specification level. Management of the roles
is anyway very practical, especially by use-cases when the project
uses different databases in the same cluster and one user should
have different rights in different databases.