Cosmetic changes in TODO.org
I suggest the following rules for maintaining TODO.org:
- Wrap all text to 80 characters (too long lines are hard to read)
- Write item titles without a trailing period (they are not a sentence)
- Make item titles brief but explanatory (similar as git commit titles)

I'd also suggest not to indent text paragraphs.  The hierarchy visually
clear enough from highlighting and folding, so I see no reason for
indent and it is harder to write.  I didn't make this change in this
commit as I don't know whether there are some other reasons for this.