Preserve virtual field values in _commit_form()
Don't reset virtual field values to their default values on commit.  The
problem appeared in InputForm where the edited values of virtual fields
didn't get to the result row, but the effect of resetting virtual fields
must be undesired in other forms to, so we tried to fix the behavior
generally in the base class.