Don't share RPyC server instances
We encounter many problems when we try to share a unique RPyC server instance
for multiple pytis2go instances. Mainly we experience the application being
completely frozen when some other pytis2go instance had to restart RPyC server
instance because of some communication problem.
So maybe it is better for now to start the own RPyC instance for each
pytis2go instance. It should not require too much resource and it should
be more safe.
Let's also increase the number of reverse forwarding ports on the application
server.