Fix scrolled query fields Apply button visibility
The button was not visible when the form was too narrow.  This fixes the
problem by creating the button inside the ScrolledWindow of the query
fields form main layout.  As this must be done in the QueryFieldsForm
class instead of the ListForm class, the change required a little bit of
refactoring, but this is ok as the code belongs more logically into
QueryFieldsForm anyway.