Merge CalendarView into Pytis JS/CSS files
Don't rely on calendarview.js and calendarview.css as external
dependency, but rather include them in base JS/CSS files.  This prevents
their unavailability on async requests (when the initial form doesn't
load them, but the later asynchronously loaded content does).  This is
actually a preparation for inline editation support.