Keep busy cursor while locking row for update
Locking may take a few seconds on some views.  The user may be surprised
when opening PopupEditForm, cursor is (up to now) busy only for the time
of form initialization and it is already switched to normal cursor when
run() is called but nothing actually appears on the screen before
ShowModal() is called.  Thus it seems logical to keep the cursor busy
also during run() up to the ShowModal() call when the form finally
appears on the screen.