Fix specific computer/runtime filter interaction
When a change in one field triggers a computer and runtime filter on
another field at the same time (or in other words if both computations
of the later field depend on the same source field value), the runtime
filter must be recomputed first in order to correctly validate the new
computed value.  This change should handle this special case correctly.