Delay RPyC tunnel start until server is ready
The tunnel would not start if the server was not started yet in the time
of session initialization.  But because server start may take some time
when ports are not available (the loop needs to timeout on every
attempted port number), we need to delay tunnel startup until the server
is ready.  This commit fixes the problem by moving tunnel initialization
to the main loop...