Invoke computers only when input is valid
Don't invoke computers when the input fields (on which the computer
functiuon depends) are not strictly valid.  Use the previous value
instead of the computed value until a new value can be computed.

This change replaces the previous attempt to handle invalid input
individually for each computer through the 'fallback' argument.  This
approach was, however, limited to defining a particular fallback value
and didn't allow to simply use the previous value.