Handle possible DBException in counting thread
This can occur when we use table functions (e.g. plpgsql function)
containing some error. This error was not caught and the user interface
has stayed frozen. So we have to catch this database exception
and propagate it to the main thread in order to show the usual
bug report dialog for the user.