Avoid conversion when validating computer deps
Validation consist of two phases - converting a string into a Value
instance and checking constraints for this instance.  As we only had one
method - 'Type.validate()' which either does both pases or only the
first one (when 'strict' is False), we had no way to do the second phase
only.  This commit fixes the problem by defining a new method
'Type.check_constraints()' which does the second phase only.

This now allows checking constraints for computer dependencies without
converting the field value to a string and back.