Remove method PresentedRow.original_presented_row()
When attempting to write a test for this method, I discovered that its behavior is chaotic.  The
returned row doesn't match the original row after its initialization, because prefill is not
re-applied.  Also non-computed virtual fields are not reset.  As the method doesn't seem to be used
anywhere, it removing it seems easier than attempting to fix it.