Avoid recursive computer input validation
Offer solution for situations where cyclic dependencies cause recursion
when validating computer input.  The recursion happens when a computer
depends on a field, which has a runtime filter or runtime arguments
which depend on the original field.  In this case the inputs of the
original field's computer funmction can not be validated, because the
validation depends on the result of the computer function.  The
recursion can be avoided by excluding given field from
validation (running the computer function with a possibly invalid input
value).  This may comlicate the logic of the computer function (may not
rely on validation) but makes the computation at least possible.  The
field can be excluded from validation explicitly by including it in the
new 'novalidate' argument of 'Computer' specification.