Cache PresentedRow field validation results
This should save multiple passes of validation for the same field
value.  Field validation often consists of DB requests and it is
now invoked several times due to the computer input validation.
The cache now keeps the result only for the last field value which
should be sufficient for the purpose.