Avoid common setup for all tests
Some tests don't need the common initializations, so they should not be
performed generally.  The tests which use the common field
specifications now use the method _mega_row().  Using common fields
proved to be problematic so it should be avoided in new tests and if
possible also ocassionaly droped from the existing tests.