Avoid redundant select on form startup
The redundant select was caused by application of the initial profile in
_on_idle.  The initial profile was applied in _on_idle only because of
the need to warn the user in case of an invalid saved profile (not
matching the potentially changed specification).

We avoid the additional select by applying the initial profile
immediately in constructor.  In case of an invalid profile, we use the
default profile instead and mark the problem to be able to warn the user
later in _on_idle.  This seems to solve the situation and be overall
more logical and simple.