Add trailing newline to EOF
Some unix tools (and POSIX spec) understand text files as sequence of
lines, where a line is a sequence of characters ending with a
newline. Therefore the last line has to end with a newline, or it
would not be a line and therefore the file would not be a sequence of
lines and therefore not a text file.

This is a problem e.g. with sed, which during interactive edits, even
without doing any substitutions, adds newlines at the end of files not
ending with newline.