Run client session in a thread
Running threads in a wx application doesn't work when x2go is imported.
This was the reason for splitting the modules and rearranging the code
in order to avoid x2go imports in any code which runs within the process
of the wx application.  This also required to isolate broker
communication into another separate process (runbroker.py).