Manage DISPLAY for client processes
This commit fixes running client processes on windows with the globally
started X-server by setting their DISPLAY environment variable according
to the started X-server.

At the same time it is now possible to specify the server variant when
starting the server and we are prepared for the option of starting
specific variants for specific session parameters.  We just need to
decide how the parameters map to server variants and manage the running
servers.