Fix session access in session manager
This required a decent redesign of the ListCtrl handling as we need to be able
to load the list multiple times when the session manager frame is now reused.
We also need a simple and straightforward mapping of session instances and list
ctrl rows, but ListCtrl.SetItemData() only allows numeric values.