Do not ignore prefill when using inserted_data
When we try to import CSV data into the side form, we usually don't have binding
column in the layout of EditForm, because it is automatically prefilled
according to the value in the main form.
In such a case the problem occurs by saving the record to database.
So we try to use the prefill like usually and only after initializing
the row with prefill we try to fill also all other values.