Prefil DB password by X2Go session password
Attempt to use the X2Go session password to log in to the database within the
application, assuming the passwords may be synchronized (LDAP).

The password is currently ignored when the configuration option
'login_selection' is used.  We also don't check whether the database user name
and the session user name match.  The password is only used if password
authentication was used to log in to the application server from Pytis2go (thus
not available when public key authentication is used, for example).