Test pytisproc version before trying to call new methods
This should avoid problems when the client contains an older pytisproc version
and the server runs a more recent version of pytis.remote.  In this case the
try/except block would handle it anyway, but it still seems safer to test and
keep it as an example for future API changes which might not suffice with
try/except.