Fix bug in instatiating SQLObjects with the same pytis name
We can have different SQLObjects sharing the same pytis name. This is usually
the case when we define the table or view with the same database name but in
different database schemes.

This situation was previously not considered because there was a dictionary of
init_functions where we could register only one function for one pytis name. Now
we register all init functions for the specific pytis name so each specification
class will be instantiated properly.