Fix increamental search behavior in dual form
The SetFocus() call causes all text field content to be selected during typing
in the search field which makes it impossible to type more than one letter.  I
am not sure why exactly is the focus restored here, but appearantly, this
change should not harm but it still helps to solve the problem described above.