Handle temporary print file removal properly
The external viewer of the generated PDF file is started
in a separate process and works with the given name
of temporary file. This file is created in the main
another process and when finished the temporary
file was removed before the viewer process could work
with it. So we have to handle the temporary file removal
ourselves and this should be done after finishing
the external viewer program.