Pass arguments to DescriptiveDualForm side form
When opening DescriptiveDualForm on a specification with a table function which
needs arguments, these arguments are passed to DescriptiveDualForm within
kwargs.  These are already passed correctly to the main form, but the side form
needs them too.

Without passing the arguments on to the side form, an error occures when
opening DescriptiveDualForm from a multi form side form, which computes its
arguments from the main form row through a function defined in
'Binding.arguments()'.