Save & restore main application frame size
The previous attempt (the commit e88d290bb reverted by 15b23638a) moved parts
of application initialization before database login, so these initializations
stopped to work in case that the database access needs the username and
password to be given interactively.  This version of the commit leaves the
initializations intact and only sets the frame size later after its
creation (which is still before it is displayed).