Avoid empty radio item for boolean fields
The boolean type defines the default value as False without the need to define
this default in specification.  Thus the radio button needs to check the type's
default value in addition to specification defined default value in order to
decide properly on the existence of the null value item.  In result, this fixes
the appearance of not null boolean radio fields.