Confirm restored forms on startup
It makes more sense to select the forms to be restored on startup than on
exit (on exit I don't know which forms I will want to restore on next startup).

Morover, it sometimes happens that some form causes troubles and the
application can not be started with it.  This had to be solved by manually
deleting the saved forms from the database, which is not practical at all.