Fix INSERT into some VIEWs with serial key
When a VIEW has a INSERT RULE without RETURNING part, and a serial
key, the code is supposed to figure out inserted row keys by querying
the database. But this code was not hit by unit tests.

So we are adding a unit test, and fixing a bug discovered by it. It
seems this part of code is not used anywhere, as previously the code
path always threw an exception due to insufficient _Query.update
args.