Adjust clipboard callback setup for GTK+ 3
The original solution was specific for GTK+ 2.  This new solution is specific
to GTK+ 3.  Unfortunetely we don't know of a way to make this code universal as
the attempt to import gtk/gi completely freezes the program when run in the
other environment.  This is not a problem as long as the wx 3.0 code (current
master) is run on GTK+ 2 only (which it typically is) and the wx 4.0 code (this
branch) is run on GTK+ 3 (which is always the case).