Update also config.dbuser after DB login
When the DB user is changed in the login dialog config.dbconnection was
updated, but config.dbuser was left with the default value (the current system
user).  This is dangerous, because dbuser is used at some places.  Actually,
config.dbuser should not be used for anything else than initializing
config.dbconnection, but as it may not be always obvious and the resulting
problem is very hard to detect, it seems safer to update dbuser too.