Move help icon when Choice size changes
When the Choice field enlarges due to a runtime filter change, the help icon
used to stay at the same place without this change.  Thus the enlarged field
overlapped the icon and it was not possible to display the help any more (even
though the icon was visually above the field).