Remove code moved to Pytis2Go
Pytis2go is not Pytis any more so the library should not share the same name.
Thus 'pytis.x2goclient' and parts of 'pytis.remote' (simply everything which
runs on the client side) are now moved to the new python library 'p2go'.  The
only code common to Pytis and Pytis2go is the RPyC service authenticator.  It's
client side must be duplicated (not shared) in Pytis in order to avoid runtime
dependencies between Pytis and Pytis2Go.