Bug report dialog improvements
The dialog now allows writing user message with more information about the
error and sending the bug report by email before closing the dialog, which
makes it possible to decide to exit/ignore independently after the bug report
is sent.  The displayed thraceback is now not visible at first as it is useless
for an ordinary user anyway.  On the other hand, when displayed, it uses the
HTML format which is a little more legible than the plain text traceback.