Disable inline editation
Inline editation needs to be completely reworked.  It doesn't work at all so it
is useless to leave it enabled.  The confusing code is now removed from
inputfield.py, but most of the code remains in _grid.py and list.py.  It should
be eventually removed frome there too or rewritten if the new concept would be
close to the original.