Remove input dialogs keeping a lagacy hack
The input dialogs are removed, but run_dialog now includes a hack transofming
their call to an InputForm call.  All (and only those) arguments which were
actually found to be in use by applications are supported by this hack.

In any case InputDialog, InputNumeric and InputDate are now deprecated and the
new functions 'input_text', 'input_date' and 'input_number' should be used
instead.