Use 'bytes' instances for internal values of Binary type
We need to use 'bytes' instead of 'buffer' in order to be future compatible
with Python 3.  We also want to avoid complicated handling of binary values
previously caused by wrapping the actual binary value inside the Binary.Buffer
instance instead of letting the internal value directly contain the binary
data.

Put simply: We want row['data'].value() to directly be a 'bytes' instance
(as with the other types) instead of having to access it through
row['data'].value().buffer().

This commit makes this possible but it still keeps the buffer() method for
backwards compatibility.  Also the Binary.Buffer (and derived classes) still
exist even though they are renamed to Binary.Data (Buffer would be confusing
now) and should not be accessed directly any more.