Don't use Binary.Data.buffer() any more
The intention of the Binary type rewrite actually was to get rid of the need to
access the binary data through the buffer() method and have the data directly
in the value representation.  The buffer() method is now still kept for
backwards compatibility, but it should not be used.  This commit removes it
throughout Pytis.