Avoid IndexError on AuiNotebook initialization
We received tracebacks where the first removed line of this commit raised
IndexError - no AuiTabCtrl children were found.  This way it should be ok.  It
is unclear why EVT_AUINOTEBOOK_BEGIN_DRAG was also bound to tabctrl, but it
seems to work fine with nb directly and it is also the documented approach.