Prevent printing huge binary data in logs
File contents currently gets logged for example in COMMAND_HANDLED_ACTION "Open
file <filename>" (see 'file_menu_items()' in screen.py).  It is also
questionable whether we want to log command arguments in 'Command.invoke()' for
security reasons, but this is not the objective of this commit.  This commit
only aims to prevent huge data dumps in log files.