Move buffering logic into the FetchBuffer class
Filling the buffer from its user side is problematic, because it requires
access to its internals and thus changing buffering strategies is complicated
and the code is hard to read.  This commit moves the buffer filling logic into
the class so that the buffer itself decides when to fill and which range to ask
for.

This change also allows implementation of a more efficient filling strategy.
Now the buffer maintains a continuous cache content in much more situations.
The buffered chunks are always designed to adjoin to the current buffer content
from both sides if possible and the buffer is only discarded if the newly
requsted item isn't located within the reach of 'fetch_size' from the current
buffer content.  Also when starting again, the buffer is filled in both
directions from the requested item, so that the requested item lies in the
middle of the newly buffered area and the buffer can then grow into both sides
with the same efficiency.