Allow fetching data rows by absolute position
To maintain backwards compatibility and still allow changing the the arguments,
the new method 'fetch()' is now the preferred interface for fetching rows and
'fetchone()' is now deprecated, only kept for backwards compatibility with its
original arguments.  This also allows removing the 'transaction' argument which
was long deprecated so it now doesn't complicate the buffer API.

'fetch()' now has a single argument 'position' which can be either a
direction (as before) or a number denoting the absolute position of the row to
fetch.  This should greatly simplify some use cases.