Simplify grid row fetching using absolute positions
Make use of fetching by absolute position because this is what the grid table
actually needs and translation of the absolute positioning into relative and
back is thus an unnecessary hassle.  This avoids most of the quite complicated
logic for retrieving rows from the data object.  We also don't need hacks
attempting to influence buffer filling strategy as the new buffer already does
that quite well.