Use cachetools.LRUCache in grid table
Try caching individual rows in a LRU cache instead of caching a continuous
areas of adjacent rows.  This might improve performance in combination with the
reimplemented buffer.  We should control the buffer size and cache size by two
distinc configuration options, but for trying it out we just use the same
option now.