Include secret column styles in cache
The cache now includes the value anyway, it is just unused for secret columns.
The new solution seems more logical as the resulting styles are all determined
at one place.  Moreover it should be more performant because it saves secret
columns lookup for each cell.  It is true, that the previous solution saved
cache lookup in case of a secret column, but considering that secret columns
are quite rare, it is better to save unnecessary secret column lookup.