Attempt to optimize grouping computation.
This change tries to eliminate repetitive computation of the nearest cached
row.  It is quite expensive and doesn't change through the recursive calls.
Also the cache lookup is unnecessary during the recursive calls (we know the
neighbour rows are not there), so it is possible to recourse only within the
inner code which is now separated into the get_group() function.