Cache grid table columns by indexes rather than by keys
This should improve grid performance as we completely avoid the translation of
keys to indexes and vice versa.  Also the lookup itself is slightly faster with
indexing and it is completely natural in case of a grid table so it doesn't
actually add any complexity or harm readability.  Last but not least, caching
lists is more memory efficient than caching dictionaries.