Avoid unnecessary data.select() in file_menu_items
We create a new data object for the Record instance in order to make a select
for finding out if the binary field is empty or not.

Because of this it is not necessary to call also initial select for the newly
created data object. Moreover sometimes this initial select can be very time
consuming, especially with very complex forms. So we simply avoid calling
this unnecessary initial select.