Optimize nearest cached group computation
This optimization is based on benchmarking previous commits and tries to fix
the worsened results between commits 7d6ff5c and 8c7970a.  It was caused by
computing min from a large set, while we ofthen only need to filter the set
first.  This optimization has a little effect in the current code, because we
now don't compute the nearest row repeatedly for neighbour rows, but it still
makes sense to apply this change.