Make fetch() argument mandatory and resurrect fetchone()
We will keep fetchone() as a convenience alias for 'fetch(FORWARD)' (without
the possibility to fetchone(BACKWARD) as this is never used in existing
application code).  This also makes it possible to make the fetch() argument
mandatory, which is definitely better because fetch() without arguments doesn't
clearly express what it does.