Remove the unused Pipe from util
There has been a bug in read() since commit 1ac5ba8d "Replace with_lock by a
context manager" because return was incorrectly replaced by break in an inner
loop.  This made the unit tests fail, but it is not worth fixing because it is
not used anywhere.  Thus simply removing.