Work around unicode builtin compatibility
This is an intermediate solution.  It gets rid of using the 'unicode' built in
type/function.  The Python 2 functionality remains unchanged and degrades to
just 'str' on Python 3.  The function call will be useless on most places once
Python 2 support is not needed.  Some places will need to be further refactored
to work well on Python 3.  But the most important thing for now is that it
doesn't change anything for Pythpon 2 and that we have all these places marked
for further hacking and we don't mix them yet with places where str was
previously used for some reason (which would happen if we imorted str from
futurize's builtins.