Make classes in caching.py Python 2/3 compatible
super() doesn't work with old style classes, but collections.UserDict is an old
style class in Python 2.  It was before solved by defining _Cache with base
classes (object, collections.UserDict), but this doesn't work in Python 3, so
the old 'super_' hack comes to rescue here until we are able to drop Python 2
support.